+++ src/main/kotlin/dev/nyon/magnetic/config/Config.kt
@@ -1,8 +1,14 @@
 package dev.nyon.magnetic.config
 
+import dev.nyon.konfig.config.config
+import dev.nyon.konfig.config.loadConfig
 import kotlinx.serialization.Serializable
+import net.fabricmc.loader.api.FabricLoader
 
-lateinit var config: Config
+val config: Config by lazy {
+    config(FabricLoader.getInstance().configDir.resolve("magnetic.json"), 1, Config()) { _, _ -> null }
+    loadConfig()
+}
 
 @Serializable
 data class Config(